LayoutTests/http/tests/media/modern-media-controls/skip-back-support/skip-back-support-button-click.html

<script src="../../../../../resources/js-test-pre.js"></script>
<script src="/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/buttons-container.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/start-button.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/play-pause-button.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/skip-back-button.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/airplay-button.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/pip-button.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/fullscreen-button.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/mute-button.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/tracks-button.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/slider.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/volume-slider.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/scrubber.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/time-label.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/time-control.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/placard.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/airplay-placard.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/pip-placard.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/media-controls.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/macos-media-controls.js" type="text/javascript"></script>
<script src="/modern-media-controls/controls/macos-inline-media-controls.js" type="text/javascript"></script>
<script src="/modern-media-controls/media/media-controller-support.js" type="text/javascript"></script>
<script src="/modern-media-controls/media/elapsed-time-support.js" type="text/javascript"></script>
<script src="/modern-media-controls/media/mute-support.js" type="text/javascript"></script>
<script src="/modern-media-controls/media/playback-support.js" type="text/javascript"></script>
<script src="/modern-media-controls/media/remaining-time-support.js" type="text/javascript"></script>
<script src="/modern-media-controls/media/scrubbing-support.js" type="text/javascript"></script>
<script src="/modern-media-controls/media/skip-back-support.js" type="text/javascript"></script>
<script src="/modern-media-controls/media/start-support.js" type="text/javascript"></script>
<script src="/modern-media-controls/media/volume-support.js" type="text/javascript"></script>
<script src="/modern-media-controls/media/media-controller.js" type="text/javascript"></script>
<script src="/modern-media-controls/main.js" type="text/javascript"></script>
<body>
<video src="../../resources/hls/generate-vod.php?duration=8000" style="width: 320px; height: 240px;"></video>
<div id="shadow"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>SkipBackSupport</code> behavior by clicking on the mute button.");

iconService.directoryPath = "/modern-media-controls/images";

const shadowRoot = document.querySelector("div#shadow").attachShadow({ mode: "open" });
const media = document.querySelector("video");
const mediaController = createControls(shadowRoot, media, null);

let numberOfEvents = 0;
function mediaDidSeek()
{
    debug("Media seeked to currentTime = " + media.currentTime);

    numberOfEvents++;
    if (numberOfEvents == 3) {
        debug("");
        shadowRoot.host.remove();
        media.remove();
        finishJSTest();
        return;
    }

    debug("");
    debug("Dispatching click event to the skip back button");
    mediaController.controls.skipBackButton.element.click();
}

function mediaCanPlay()
{
    media.removeEventListener("canplaythrough", mediaCanPlay);
    media.addEventListener("seeked", mediaDidSeek);
    media.currentTime = 50;
}

media.addEventListener("canplaythrough", mediaCanPlay);

</script>
<script src="../../../../../resources/js-test-post.js"></script>
</body>
