LayoutTests/http/tests/media/modern-media-controls/skip-back-support/skip-back-support-button-click-expected.txt

Testing the SkipBackSupport behavior by clicking on the mute button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Dispatching click event to the skip back button

seeked event was dispatched
PASS media.currentTime is 20

Dispatching click event to the skip back button

seeked event was dispatched
PASS media.currentTime is within 0.1 of 0

PASS successfullyParsed is true

TEST COMPLETE
